Only update feeds listed in CURRENT_FEEDS
Bulk importing arxiv data reveals a number of older categories which
seem to be partially obsolete and do not have rss endpoints. The
arxiv_update task fails if it attempts to retrieve new papers for these.
Added a global settings.rb file with a CURRENT_FEEDS value to rectify
this.