Disable regular authorship syncing for now
export.arxiv.org tends to ban our IP. Since the authorship
update requests scale linearly with the number of users,
this may be the culprit. It's more important to keep the
core functionality of receiving new papers intact until we
find a better solution for this (or at least offload it
to a secondary IP).