Do not require authentication to download JSON file with scites
The list of scites of a user is currently public, but it can be
downloaded in JSON format only by the user who made them after
authenticating. I do not see why this should be case.

This change should remove the /download_scites route and add a
/:username/download_scite, which should not require
authentication. The button "Download scites" on the user profile
should direct to this link, so that nothing changes on the UI side.

I am not proficent in Ruby and I did not check if this actually works. Sorry!