Add 'jobs' feature
This implements the ruby/rails side of the 'SciRate jobs' mini-feature.
We've got the compiled elm (under app/assets/javascripts), as well as
all the styling.

We've got the views, and the minor backend ruby changes to create the
JWTs and do emailing.

And finally there is the migration to allow users to control whether or
not they are shown jobs at all.